refactor(slice): Slice와 관련된 Entity들,  CourtTimeSlot Entity로 통합 #2
- 패키지 명: slice  -> courttimeslot
- Borrow, Match Entity -> CourtTimeSlot Entity로 통합
- CourtTimeSlot 엔티티 구현
