+++ src/main/java/com/example/tuktuk/domain/courttimeslot/CourtTimeSlotId.java
@@ -1,0 +1,15 @@
+package com.example.tuktuk.domain.courttimeslot;
+
+import jakarta.persistence.Column;
+
+import java.io.Serializable;
+
+public class CourtTimeSlotId implements Serializable {
+
+    @Column(name = "court_time_slot_id")
+    private long id;
+
+    protected CourtTimeSlotId() {
+
+    }
+}
