+++ src/main/java/com/example/tuktuk/domain/courttimeslot/Type.java
@@ -1,0 +1,14 @@
+package com.example.tuktuk.domain.courttimeslot;
+
+public enum Type {
+
+    NONE("없음"),
+    MATCH("매치"),
+    RENTAL("대여");
+
+    private String value;
+
+    Type(String value) {
+        this.value = value;
+    }
+}
