--- src/main/java/com/example/tuktuk/domain/slice/Slice.java
@@ -1,33 +1,0 @@
-package com.example.tuktuk.domain.slice;
-
-import com.example.tuktuk.domain.stadium.CourtId;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "slice")
-public class Slice {
-
-    @Id
-    @Column(name = "id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Embedded
-    private CourtId courtId;
-
-    @Embedded
-    private Time time;
-
-    @Enumerated
-    @Column(name = "type", nullable = false)
-    private SliceType sliceType;
-
-}
