--- src/main/java/com/example/tuktuk/domain/slice/SliceId.java
@@ -1,15 +1,0 @@
-package com.example.tuktuk.domain.slice;
-
-import jakarta.persistence.Column;
-
-import java.io.Serializable;
-
-public class SliceId implements Serializable {
-
-    @Column(name = "slice_id")
-    private long id;
-
-    protected SliceId() {
-
-    }
-}
