--- src/main/java/com/example/tuktuk/domain/slice/borrow/BorrowStatus.java
@@ -1,12 +1,0 @@
-package com.example.tuktuk.domain.slice.borrow;
-
-public enum BorrowStatus {
-    AVAILABLE("예약 가능"),
-    UNAVAILABLE("예약 불가");
-
-    private final String displayName;
-
-    BorrowStatus(String displayName) {
-        this.displayName = displayName;
-    }
-}
