--- src/main/java/com/example/tuktuk/domain/slice/gamematch/GameStatus.java
@@ -1,8 +1,0 @@
-package com.example.tuktuk.domain.slice.gamematch;
-
-public enum GameStatus {
-    NOT_ENOUGH_PLAYERS,
-    ENOUGH_PLAYERS,
-    FULL_CAPACITY;
-
-}
